Support "symbolic" configs.
A config entry is symbolic if it binds directly to a Python symbol, rather than
binding to a string containing the name of a Python symbol. This means that
apps using symbolic bindings find out early, as soon as the config _class_ is
defined, whether they've misspelled a class or module name.

I've also simplified the string-to-symbol resolution a bit.